chore: refactor logic from UI manager and add lookupUtils
